+++ packages/select/src/resources.d.ts
@@ -45,6 +45,20 @@ export declare const AvNavigationSelect: <
   props: PrebuiltResourceSelectProps<Option, IsMulti, Group>
 ) => JSX.Element;
 
+type Payer = {
+  id: string;
+  payerId: string;
+  payerName: string;
+};
+
+export declare const AvPayerSelect: <
+  Option = Payer,
+  IsMulti extends boolean = boolean,
+  Group extends GroupBase<Option> = GroupBase<Option>
+>(
+  props: PrebuiltResourceSelectProps<Option, IsMulti, Group> & { customerId: string }
+) => JSX.Element;
+
 type Address = {
   line1: string;
   line2: string;