+++ packages/table/src/CellDefinitions/ActionCell.tsx
@@ -15,27 +15,36 @@ const ActionCell = <T extends IdType>({
   actions,
   primaryAction,
 }: ActionCellConfig<T>): ((cell: Cell<T>) => JSX.Element) => {
-  const ActionCellDef = ({ row: { original }, index }: Cell<T>): JSX.Element => (
-    <>
-      <TableActionMenu id={`table_row_action_menu_${original.id ? original.id : index}`}>
-        {actions.map((action, index) => (
-          <TableActionMenuItem
-            key={action.id}
-            id={`table_row_action_menu_item_${index}`}
-            action={action}
-            record={original}
+  const ActionCellDef = ({ row: { original }, index }: Cell<T>): JSX.Element => {
+    const isPrimaryActionVisible = primaryAction
+      ? primaryAction.isVisible
+        ? primaryAction.isVisible(original)
+        : true
+      : false;
+
+    return (
+      <>
+        <TableActionMenu id={`table_row_action_menu_${original.id ? original.id : index}`}>
+          {actions.map((action, index) => (
+            <TableActionMenuItem
+              key={action.id}
+              id={`table_row_action_menu_item_${index}`}
+              action={action}
+              record={original}
+            />
+          ))}
+        </TableActionMenu>
+        {primaryAction && isPrimaryActionVisible && (
+          <Icon
+            data-testid={`table_row_action_menu_item_${original.id}_primaryAction`}
+            name={primaryAction.iconName}
+            title={primaryAction.title}
+            onClick={() => primaryAction.onClick(original)}
           />
-        ))}
-      </TableActionMenu>
-      {primaryAction && (
-        <Icon
-          name={primaryAction.iconName}
-          title={primaryAction.title}
-          onClick={() => primaryAction.onClick(original)}
-        />
-      )}
-    </>
-  );
+        )}
+      </>
+    );
+  };
 
   return ActionCellDef;
 };