+++ packages/table/src/types/PrimaryRecordAction.ts
@@ -2,4 +2,5 @@ export interface PrimaryRecordAction<T> {
   iconName: string;
   title: string;
   onClick: (record?: T) => void;
+  isVisible?: (record?: T) => boolean;
 }