+++ packages/table/src/types/ReactTable.ts
@@ -12,6 +12,11 @@ import {
   UseTableHeaderGroupProps,
   UseRowSelectOptions,
   UseTableOptions,
+  UseSortByOptions,
+  UsePaginationOptions,
+  UsePaginationInstanceProps,
+  UseColumnOrderInstanceProps,
+  UseRowSelectState,
 } from 'react-table';
 
 import { TableSort } from './TableSort';
@@ -25,6 +30,10 @@ export type Row<T extends IdType> = RtRow<T> & {
   original: T;
 };
 
+export type RowProps = {
+  isDisabled?: boolean;
+} & React.HTMLAttributes<HTMLTableRowElement>;
+
 export type Column<T extends IdType> = RtColumn<T> & {
   className?: string;
   stickyRight?: boolean;
@@ -35,6 +44,7 @@ export type Column<T extends IdType> = RtColumn<T> & {
   disableSortBy?: boolean;
   label?: string;
   hidden?: boolean;
+  isVisible?: boolean;
 };
 
 export type Cell<T extends IdType> = RtCell<T> & {
@@ -60,19 +70,26 @@ export type TableOptions<T extends IdType> = RtTableOptions<T> &
     autoResetSelectedRows?: boolean;
     autoResetSortBy?: boolean;
     manualSortBy?: boolean;
-  };
+    hiddenColumns?: string[];
+    initialState?: Partial<TableState<T>> & CurrentTableState<T>;
+  } & UseSortByOptions<T> &
+  UsePaginationOptions<T>;
 
 export type TableInstance<T extends IdType> = UseSortByInstanceProps<T> &
   UseRowSelectInstanceProps<T> &
   UseRowSelectOptions<T> &
-  RtTableInstance<T> & {
+  RtTableInstance<T> &
+  UseColumnOrderInstanceProps<T> &
+  UsePaginationInstanceProps<T> & {
     toggleAllPageRowsSelected: (set?: boolean) => void;
     selectedFlatRows: Row<T>[];
     isAllRowsSelected: boolean;
     toggleAllRowsSelected: () => void;
     sortBy?: TableSort;
+  } & {
+    manualSortBy?: boolean;
   };
 
-export interface CurrentTableState extends TableState {
+export type CurrentTableState<T extends IdType> =  TableState & Partial<UseRowSelectState<T>> & {
   sortBy?: TableSort[];
-}
+};