+++ packages/favorites/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 3.3.0 (2022-05-09)
+
+
+### Features
+
+* **favorites:** added remove content for tooltip and refactored tests to fix order bug and improve tests ([e58b5a0](https://github.com/Availity/availity-react/commit/e58b5a0026b406656f1e036534ae9a155a5665f8))
+
+
+
+
+
+# 3.2.0 (2022-05-03)
+
+
+### Features
+
+* **favorites:** add `disabled` prop to heart ([fb42164](https://github.com/Availity/availity-react/commit/fb42164ac413676e8c3e35484566fca0cb46a70d))
+* **favorites:** add test for `disabled` prop ([93031cb](https://github.com/Availity/availity-react/commit/93031cb07c0634023af1eb984dd3ccc917aa866d))
+
+
+
+
+
 # 3.1.0 (2022-03-08)